Renamed config and discovery packages to internal in the freebox binding. See openhab/static-code-analysis#100
